+++ src/core/object/watch/engines/proxy.ts
@@ -229,6 +229,14 @@ export function watch<T extends object>(
 
 				} else {
 					normalizedKey = key;
+
+					const
+						desc = Object.getOwnPropertyDescriptor(target, key);
+
+					// Readonly non-configurable values can't be wrapped due Proxy API limitations
+					if (desc?.writable === false && desc.configurable === false) {
+						return val;
+					}
 				}
 
 				if (propFromProto || !isArray && !Object.hasOwnProperty(target, key)) {